fix: Server Class内の各関数を機能別にファイルを作って保存した。
