+++ Makefile
@@ -1,7 +1,7 @@
 NAME = Server
 CXX = c++
 CXXFLAGS = -std=c++98 -MMD -MP -pedantic
-SRCS = Server.cpp Main.cpp ClientData.cpp ClientAuth.cpp CmdResponse.cpp Commands.cpp NICK.cpp
+SRCS = Server.cpp Main.cpp ClientData.cpp ClientAuth.cpp CmdResponse.cpp Commands.cpp NICK.cpp USER.cpp PASS.cpp Send.cpp Receive.cpp Utils.cpp
 VPATH = ./command
 DEPFILES = $(OBJS:%.o=%.d)
 OBJS = ${addprefix $(OBJS_DIR)/,$(SRCS:.cpp=.o)}
