+++ objs/PASS.d
@@ -1,0 +1,7 @@
+objs/PASS.o: command/PASS.cpp command/../Server.hpp \
+ command/../ClientData.hpp command/../Server.hpp \
+ command/../CmdResponse.hpp
+command/../Server.hpp:
+command/../ClientData.hpp:
+command/../Server.hpp:
+command/../CmdResponse.hpp:
