+++ objs/Receive.d
@@ -1,0 +1,4 @@
+objs/Receive.o: Receive.cpp Server.hpp ClientData.hpp CmdResponse.hpp
+Server.hpp:
+ClientData.hpp:
+CmdResponse.hpp:
