+++ objs/Send.d
@@ -1,0 +1,4 @@
+objs/Send.o: Send.cpp Server.hpp ClientData.hpp CmdResponse.hpp
+Server.hpp:
+ClientData.hpp:
+CmdResponse.hpp:
