+++ objs/USER.d
@@ -1,0 +1,7 @@
+objs/USER.o: command/USER.cpp command/../Server.hpp \
+ command/../ClientData.hpp command/../Server.hpp \
+ command/../CmdResponse.hpp
+command/../Server.hpp:
+command/../ClientData.hpp:
+command/../Server.hpp:
+command/../CmdResponse.hpp:
